fix wine controller detection, resize fc select windows, re-add switch script

+++ kumiteos/airootfs/etc/skel/.scripts/copy-old-fc-roms.sh
@@ -1,20 +1,23 @@
 #!/bin/bash
 FCDIR="${HOME}/.fightcade2"
-OLDFCDIR=$(yad --file --title="Select Fightcade Directory")
+OLDFCDIR=$(yad --file --title="Select Fightcade Directory" --width=800 --height=600)
 EMULATORS="fbneo
 flycast
 ggpofba
 snes9x"
-for emu in $EMULATORS
-do
-    if [ -L "${FCDIR}/emulator/${emu}/ROMs" ]
-    then
-        unlink "${FCDIR}/emulator/${emu}/ROMs"
-        mkdir "${FCDIR}/emulator/${emu}/ROMs"
-    fi
-    if [ -d "${OLDFCDIR}/emulator/${emu}/ROMs" ] 
-    then
-        cp -R "${OLDFCDIR}/emulator/${emu}/ROMs/." "${FCDIR}/emulator/${emu}/ROMs"
-    fi
-done
+if [ -d "${OLDFCDIR}" ]
+then
+    for emu in $EMULATORS
+    do
+        if [ -L "${FCDIR}/emulator/${emu}/ROMs" ]
+        then
+            unlink "${FCDIR}/emulator/${emu}/ROMs"
+            mkdir "${FCDIR}/emulator/${emu}/ROMs"
+        fi
+        if [ -d "${OLDFCDIR}/emulator/${emu}/ROMs" ]
+        then
+            cp -R "${OLDFCDIR}/emulator/${emu}/ROMs/." "${FCDIR}/emulator/${emu}/ROMs"
+        fi
+    done
+fi
